requirements.txt
email
re
sqlite3
os
python-dotenv
google-generativeai
smtplib
mimetypes
reportlab
fastapi
uvicorn
requests
beautifulsoup4
pandas
easyocr
transformers
agno
openai
pypdf
groq
lancedb
pdf2image
fitz
pymupdf